Move the repository intrefaces to the proper project
