+++ student/pom.xml
@@ -37,7 +37,6 @@
 		<student-shared.version>0.0.1-SNAPSHOT</student-shared.version>
 		<jackson.version>2.8.5</jackson.version>
 		<resource-bundler.version>0.0.1-SNAPSHOT</resource-bundler.version>
-
 	</properties>
 
 	<dependencies>
