+++ student/src/main/java/tds/student/sql/repository/ExamRepository.java
@@ -1,4 +1,4 @@
-package tds.student.sql.abstractions;
+package tds.student.sql.repository;
 
 import TDS.Shared.Exceptions.ReturnStatusException;
 import com.google.common.base.Optional;
