+++ student/src/test/java/tds/student/services/remote/RemoteAccommodationsServiceTest.java
@@ -20,8 +20,8 @@ import tds.common.Algorithm;
 import tds.exam.ApproveAccommodationsRequest;
 import tds.exam.ExamAccommodation;
 import tds.student.services.abstractions.IAccommodationsService;
-import tds.student.sql.abstractions.AssessmentRepository;
-import tds.student.sql.abstractions.ExamRepository;
+import tds.student.sql.repository.AssessmentRepository;
+import tds.student.sql.repository.ExamRepository;
 import tds.student.sql.data.AccommodationType;
 import tds.student.sql.data.AccommodationValue;
 import tds.student.sql.data.Accommodations;
